Add multiDetector to resource class
