+++ buildscripts/import-control.xml
@@ -206,7 +206,6 @@ General guidelines on imports:
     <allow pkg="io.opencensus.common"/>
     <allow pkg="io.opencensus.implcore"/>
     <allow pkg="io.opencensus.metrics"/>
-    <allow pkg="io.opencensus.resource"/>
     <allow pkg="io.opencensus.stats"/>
     <allow pkg="io.opencensus.tags"/>
     <allow pkg="io.opencensus.trace"/>
